api/requirements.txt
fastapi[standard]>=0.115.12
langchain>=0.3.21
langchain-openai>=0.3.11
arxiv>=2.1.3
pytest>=8.3.5
python-dotenv>=1.1.0
pypdf>=5.4.0
langchain-community>=0.3.20
chromadb>=0.6.3
